completed the start message for the websocket

+++ api/gameplay/update_gamesession.go
@@ -2,7 +2,7 @@ package gameplay
 
 import "database/sql"
 
-func updateGameSession(tx *sql.Tx, sessionUUID string, targetScore int, numberOfTeams int) error {
+func updateGameSession(tx *sql.Tx, sessionUUID string, targetScore float64, numberOfTeams int) error {
 	_, err := tx.Exec(`
         UPDATE game_sessions
         SET target_score = $1, number_of_teams = $2
